URF-28: Implemented serialization of various resource references using labels.
This implementation still refers to "idents"; this will be renamed
back to "labels".

Circular references have temporarily been removed from the tests.
It appears that both the parser and serializer support them.
Comparing the graphs for verification is not straightforward, however,
and will require a bit of work. This is a lower priority at the moment.